+++ opencti-platform/opencti-graphql/src/domain/file.js
@@ -1,5 +1,5 @@
 import * as R from 'ramda';
-import { loadFile, upload } from '../database/object-storage';
+import { loadFile, upload } from '../database/file-storage';
 import { pushToConnector } from '../database/rabbitmq';
 import { createWork } from './work';
 import { logApp } from '../config/conf';
