+++ opencti-platform/opencti-graphql/src/initialization.js
@@ -4,7 +4,7 @@ import semver from 'semver';
 import { booleanConf, logApp, PLATFORM_VERSION } from './config/conf';
 import { elCreateIndexes, elIndexExists, searchEngineInit } from './database/engine';
 import { initializeAdminUser } from './config/providers';
-import { initializeBucket, isStorageAlive } from './database/object-storage';
+import { initializeBucket, isStorageAlive } from './database/file-storage';
 import { rabbitMQIsAlive } from './database/rabbitmq';
 import { addMarkingDefinition } from './domain/markingDefinition';
 import { addSettings } from './domain/settings';
