+++ opencti-platform/opencti-graphql/src/resolvers/externalReference.js
@@ -18,7 +18,7 @@ import withCancel from '../graphql/subscriptionWrapper';
 import { RELATION_EXTERNAL_REFERENCE } from '../schema/stixMetaRelationship';
 import { buildRefRelationKey } from '../schema/general';
 import { worksForSource } from '../domain/work';
-import { filesListing } from '../database/object-storage';
+import { filesListing } from '../database/file-storage';
 import { stixCoreObjectImportPush } from '../domain/stixCoreObject';
 import { connectorsForEnrichment } from '../database/repository';
