+++ opencti-platform/opencti-graphql/src/resolvers/stixDomainObject.js
@@ -25,7 +25,7 @@ import {
 import { findById as findStatusById, getTypeStatuses } from '../domain/status';
 import { pubsub } from '../database/redis';
 import withCancel from '../graphql/subscriptionWrapper';
-import { filesListing } from '../database/object-storage';
+import { filesListing } from '../database/file-storage';
 import { ABSTRACT_STIX_DOMAIN_OBJECT } from '../schema/general';
 import { stixDomainObjectOptions } from '../schema/stixDomainObject';
 import { stixCoreObjectImportPush } from '../domain/stixCoreObject';
